requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
flask_security
gunicorn
pyyaml

